Add Learner entity and endpoints with tests
