+++ HomeschoolPlanner.Api/Data/AppDbContext.cs
@@ -44,14 +44,6 @@ public class User
     public DateTime CreatedUtc { get; set; } = DateTime.UtcNow;
 }
 
-public class Learner
-{
-    public Guid Id { get; set; }
-    public Guid UserId { get; set; }
-    public string Name { get; set; } = "";
-    public string Grade { get; set; } = "";
-}
-
 public class Subject
 {
     public Guid Id { get; set; }
