+++ HomeschoolPlanner.Api/Endpoints/Learners.cs
@@ -22,26 +22,6 @@ namespace HomeschoolPlanner.Api.Endpoints
             return Results.Created($"/api/v1/learners/{dto.Id}", dto);
         }
 
-        public static async Task<IResult> Get(AppDbContext db, Guid id) =>
-            await db.Learners.FindAsync(id) is { } l ? Results.Ok(l) : Results.NotFound();
-
-        public static async Task<IResult> Update(AppDbContext db, Guid id, Learner input)
-        {
-            var e = await db.Learners.FindAsync(id);
-            if (e is null) return Results.NotFound();
-            e.Name = input.Name; e.Grade = input.Grade;
-            await db.SaveChangesAsync();
-            return Results.NoContent();
-        }
-
-        public static async Task<IResult> Delete(AppDbContext db, Guid id)
-        {
-            var e = await db.Learners.FindAsync(id);
-            if (e is null) return Results.NotFound();
-            db.Remove(e); await db.SaveChangesAsync();
-            return Results.NoContent();
-        }
-
         private static Guid GetUserId(HttpContext ctx)
             => Guid.Parse("11111111-1111-1111-1111-111111111111"); // TODO: extract from JWT for MVP
     }
