+++ HomeschoolPlanner.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HomeschoolPlanner.Api", "Ho
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "HomeschoolPlanner.Domain", "HomeschoolPlanner.Domain\HomeschoolPlanner.Domain.fsproj", "{33844860-DCB7-46CC-9AB3-87E939674A43}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HomeschoolPlanner.Api.Tests", "HomeschoolPlanner.Api.Tests\HomeschoolPlanner.Api.Tests.csproj", "{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,18 @@ Global
 		{33844860-DCB7-46CC-9AB3-87E939674A43}.Release|x64.Build.0 = Release|Any CPU
 		{33844860-DCB7-46CC-9AB3-87E939674A43}.Release|x86.ActiveCfg = Release|Any CPU
 		{33844860-DCB7-46CC-9AB3-87E939674A43}.Release|x86.Build.0 = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|x64.Build.0 = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Debug|x86.Build.0 = Debug|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|x64.ActiveCfg = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|x64.Build.0 = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|x86.ActiveCfg = Release|Any CPU
+		{44988EC9-DB5F-489D-ACFF-3F7DAFBCC43E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
